created reusable table day sidebar
