+++ frontend/src/components/EditCurrentSchedule/Schedule/Schedule.js
@@ -2,12 +2,14 @@ import React, { Fragment, useEffect, useState } from 'react';
 import i18n from 'i18next';
 import ScheduleBoard from '../../../containers/EditCurrentSchedule/ScheduleBoard';
 import ScheduleDialog from '../../../containers/Dialogs/ScheduleDialog';
+import ScheduleDaySidebar from '../../ScheduleTable/ScheduleDaySidebar/ScheduleDaySidebar';
 
 import {
     NO_CURRENT_SEMESTER,
     COMMON_GROUP_TITLE,
 } from '../../../constants/translationLabels/common';
 import { actionType } from '../../../constants/actionTypes';
+import { addClassDayBoard, removeClassDayBoard } from '../../../helper/schedule';
 import './Schedule.scss';
 
 const Schedule = (props) => {
@@ -24,7 +26,6 @@ const Schedule = (props) => {
     } = props;
     const [isOpenScheduleDialog, setIsOpenScheduleDialog] = useState(false);
     const [dialogScheduleData, setDialogScheduleData] = useState(null);
-    const hoverLineClassName = 'hover-line';
     const days = currentSemester.semester_days;
     const classes = currentSemester.semester_classes;
 
@@ -56,25 +57,6 @@ const Schedule = (props) => {
         setIsOpenScheduleDialog(false);
     };
 
-    const getDayColour = (index) => {
-        return index % 2 ? 'dark-blue-day' : 'blue-day';
-    };
-
-    const addClassDayBoard = (dayName, classId) => {
-        const dayClassWeek = document.querySelectorAll(`.${dayName}-${classId}`);
-        const dayClass = document.querySelector(`#${dayName}-${classId}`);
-        dayClass.classList.add('focus-class');
-        dayClassWeek[0].classList.add(hoverLineClassName);
-        dayClassWeek[1].classList.add(hoverLineClassName);
-    };
-    const removeClassDayBoard = (dayName, classId) => {
-        const dayClassWeek = document.querySelectorAll(`.${dayName}-${classId}`);
-        const dayClass = document.querySelector(`#${dayName}-${classId}`);
-        dayClass.classList.remove('focus-class');
-        dayClassWeek[0].classList.remove(hoverLineClassName);
-        dayClassWeek[1].classList.remove(hoverLineClassName);
-    };
-
     return (
         <>
             {isOpenScheduleDialog && (
@@ -87,36 +69,11 @@ const Schedule = (props) => {
             )}
             {currentSemester.id ? (
                 <>
-                    <aside className="day-classes-aside">
-                        <div className="card group-title">{i18n.t(COMMON_GROUP_TITLE)}</div>
-                        {days.map((day, index) => (
-                            <div className="cards-container day-container" key={day}>
-                                <span className={`${getDayColour(index)} schedule-day card`}>
-                                    {i18n.t(`day_of_week_${day}`)}
-                                </span>
-                                <div className="class-section">
-                                    {classes.map((classScheduler) => (
-                                        <Fragment key={classScheduler.id}>
-                                            <div className="day-section">
-                                                <p
-                                                    className={`day-line ${day}-${classScheduler.id}`}
-                                                ></p>
-                                                <span
-                                                    id={`${day}-${classScheduler.id}`}
-                                                    className="card schedule-class"
-                                                >
-                                                    {classScheduler.class_name}
-                                                </span>
-                                                <p
-                                                    className={`day-line ${day}-${classScheduler.id}`}
-                                                ></p>
-                                            </div>
-                                        </Fragment>
-                                    ))}
-                                </div>
-                            </div>
-                        ))}
-                    </aside>
+                    <ScheduleDaySidebar
+                        days={days}
+                        title={i18n.t(COMMON_GROUP_TITLE)}
+                        classes={classes}
+                    />
                     <section className="groups-section ">
                         {groups.map((group) => {
                             const isSelectedGroup = group.id === groupId;
