+++ frontend/src/components/EditCurrentSchedule/Schedule/Schedule.scss
@@ -30,56 +30,6 @@
         margin: 0;
     }
 
-    .day-classes-aside {
-        position: sticky;
-        left: 0;
-        z-index: 10;
-        padding-left: 15px;
-        padding-top: 15px;
-        background-color: $snowy-white;
-        .card.empty-card {
-            width: 120px;
-            height: 19px;
-            font-size: 13px;
-            text-align: center;
-        }
-
-        .day-container {
-            position: relative;
-            display: flex;
-            margin: 0;
-            place-items: stretch;
-            .schedule-day {
-                position: relative;
-                font-size: 17px;
-                writing-mode: vertical-rl;
-                text-orientation: upright;
-                text-align: center;
-            }
-            .day-section {
-                position: relative;
-                height: 228px;
-            }
-            .schedule-class {
-                height: 228px;
-                text-align: center;
-                font-size: $font-base;
-                writing-mode: vertical-rl;
-                text-orientation: upright;
-                margin-bottom: 0.8px;
-                margin: 0 5px;
-            }
-        }
-        .dark-blue-day {
-            background-color: $light-blue;
-            color: $snowy-white;
-        }
-
-        .blue-day {
-            background-color: $blue;
-            color: $snowy-white;
-        }
-    }
     .group-title {
         display: flex;
         height: 34px;
