+++ frontend/src/components/ScheduleTable/ScheduleDaySidebar/ScheduleDaySidebar.scss
@@ -1,0 +1,61 @@
+@import '../../../styles/colors.scss';
+@import '../../../styles/variables.scss';
+
+.day-classes-aside {
+    position: sticky;
+    left: 0;
+    z-index: 10;
+    padding-left: 15px;
+    padding-top: 15px;
+    background-color: $snowy-white;
+    .card.empty-card {
+        width: 120px;
+        height: 19px;
+        font-size: 13px;
+        text-align: center;
+    }
+
+    .day-container {
+        position: relative;
+        display: flex;
+        margin: 0;
+        place-items: stretch;
+        .schedule-day {
+            position: relative;
+            font-size: 17px;
+            writing-mode: vertical-rl;
+            text-orientation: upright;
+            text-align: center;
+        }
+        .day-section {
+            position: relative;
+            height: 228px;
+        }
+        .schedule-class {
+            height: 228px;
+            text-align: center;
+            font-size: $font-base;
+            writing-mode: vertical-rl;
+            text-orientation: upright;
+            margin-bottom: 0.8px;
+            margin: 0 5px;
+        }
+    }
+    .dark-blue-day {
+        background-color: $light-blue;
+        color: $snowy-white;
+    }
+
+    .blue-day {
+        background-color: $blue;
+        color: $snowy-white;
+    }
+    .day-sidebar-title {
+        display: flex;
+        height: 34px;
+        padding: 10px 7px;
+        justify-content: center;
+        font-size: $font-sm;
+        overflow: hidden;
+    }
+}
