+++ frontend/src/helper/schedule.js
@@ -1,0 +1,37 @@
+import { isEmpty } from 'lodash';
+
+const hoverLineClassName = 'hover-line';
+
+export const divideLessonsByOneHourLesson = (items, lessons) => {
+    const lessonItem = [];
+    lessons.forEach((lesson) => {
+        let filteredLesson = [];
+        const modifiedLesson = lesson;
+        const { hours } = modifiedLesson;
+
+        if (!isEmpty(items)) {
+            filteredLesson = items.filter((item) => item.lesson.id === lesson.id);
+        }
+
+        for (let i = 0; i < hours - filteredLesson.length; i += 1) {
+            lessonItem.push(modifiedLesson);
+        }
+    });
+    return lessonItem;
+};
+
+export const addClassDayBoard = (dayName, classId) => {
+    const dayClassWeek = document.querySelectorAll(`.${dayName}-${classId}`);
+    const dayClass = document.querySelector(`#${dayName}-${classId}`);
+    dayClass.classList.add('focus-class');
+    dayClassWeek[0].classList.add(hoverLineClassName);
+    dayClassWeek[1].classList.add(hoverLineClassName);
+};
+
+export const removeClassDayBoard = (dayName, classId) => {
+    const dayClassWeek = document.querySelectorAll(`.${dayName}-${classId}`);
+    const dayClass = document.querySelector(`#${dayName}-${classId}`);
+    dayClass.classList.remove('focus-class');
+    dayClassWeek[0].classList.remove(hoverLineClassName);
+    dayClassWeek[1].classList.remove(hoverLineClassName);
+};
